astra-backend/astraweb/requirements.txt
amqp==2.2.2
astroid==1.6.2
awsebcli==3.12.4
Babel==2.5.3
billiard==3.5.0.3
blessed==1.14.2
boto3==1.7.4
botocore==1.10.4
celery==4.1.0
cement==2.8.2
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.7
Django==2.0.3
django-celery-beat==1.1.1
django-phonenumber-field==2.0.0
django-storages==1.6.6
djangorestframework==3.7.7
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
idna==2.6
isort==4.3.4
jmespath==0.9.3
kombu==4.1.0
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
passlib==1.7.1
pathspec==0.5.5
pbr==3.1.1
phonenumberslite==8.9.1
Pillow==5.0.0
psycopg2==2.7.4
pubnub==4.0.13
pycryptodomex==3.5.1
pylint==1.8.3
python-crontab==2.2.8
python-dateutil==2.6.1
python-magic==0.4.15
pytz==2018.3
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.13
semantic-version==2.5.0
six==1.11.0
stevedore==1.28.0
tabulate==0.7.5
termcolor==1.1.0
urllib3==1.22
vine==1.1.4
virtualenv==15.1.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
wcwidth==0.1.7
websocket-client==0.47.0
wrapt==1.10.11
